+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## Changelog
+version 5.0.1
+- The Plugin functions updated to fit Matomo 5.x 
+- libs/js/ClickHeat.js optimisation
 
-Here goes the changelog text.